backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
pillow
easyocr
deep-translator
python-dotenv
numpy
